setup example app
