--- example-app/src/main/res/raw/auth_config.json
@@ -1,6 +1,0 @@
-{
-  "client_id": "1079849396355-q687vpf16ovveafo6robcgi1kaoaem3e.apps.googleusercontent.com",
-  "redirect_uri": "com.googleusercontent.apps.1079849396355-q687vpf16ovveafo6robcgi1kaoaem3e:/oauth2redirect",
-  "authorization_scope": "openid email profile",
-  "discovery_uri": "https://accounts.google.com/.well-known/openid-configuration"
-}
