+++ example/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Thu Apr 18 08:02:42 BRT 2024
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-8.2-bin.zip
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
